# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ForgotPasswordRequest:
    properties:
      email:
        description: User's registered email address
        type: string
    required:
    - email
    type: object
  handler.GoogleLoginRequest:
    properties:
      id_token:
        description: Google ID token from frontend OAuth flow
        type: string
    required:
    - id_token
    type: object
  handler.RefreshTokenRequest:
    properties:
      refresh_token:
        description: Valid refresh token to exchange for new access token
        type: string
    required:
    - refresh_token
    type: object
  handler.ResetPasswordRequest:
    properties:
      code:
        description: OTP code received via email
        type: string
      email:
        description: User's registered email address
        type: string
      new_password:
        description: New password (minimum 8 characters)
        minLength: 8
        type: string
    required:
    - code
    - email
    - new_password
    type: object
...
  handler.Verify2FARequest:
    properties:
      code:
        description: 2FA verification code
        type: string
      email:
        description: User's email address
        type: string
    required:
    - code
    - email
    type: object
...
  models.LoginRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RegisterRequest:
    properties:
      email:
        maxLength: 50
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      password:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  response.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/response.UserResponse'
    type: object
  response.RegisterResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/response.UserResponse'
    type: object
  response.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@authentio.com
    name: Authentio Support
    url: https://authentio.com/support
  description: Secure authentication service with JWT, OAuth2, 2FA, and password reset
    functionality
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Authentio API
  version: "1.0"
paths:
...
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      description: Verify the 2FA code sent to user's email during login process
      parameters:
      - description: 2FA verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA verification successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or expired 2FA code
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify two-factor authentication code
      tags:
      - authentication
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset code to the user's email address
      parameters:
      - description: Password reset request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid email format
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset
      tags:
      - authentication
  /auth/google/callback:
    get:
      description: Handle Google OAuth callback with authorization code (server-side
        flow)
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OAuth authentication successful
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Missing authorization code
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Failed to exchange code for tokens
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth callback handler
      tags:
      - authentication
  /auth/google/login:
    post:
      consumes:
      - application/json
      description: Authenticate user using Google OAuth ID token (frontend flow)
      parameters:
      - description: Google ID token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Google authentication successful
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Invalid ID token format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid Google token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth login with ID token
      tags:
      - authentication
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT tokens
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT tokens
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Invalid input data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens generated successfully
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Invalid or expired refresh token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and personal information
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/response.RegisterResponse'
        "400":
          description: Invalid input data or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using verification code received via email
      parameters:
      - description: Password reset confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid code, email, or password requirements not met
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset user password
      tags:
      - authentication
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {your_jwt_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"